refactor and add arrow up and down functionality for easy navigation between slides
